flight-cancellation-composite/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
kafka-python==2.0.2
python-dotenv==1.0.0
